# chore: optimize termux_step_get_dependencies