Removed external commands invocations where possible.
United some conditions under single test invocation
Refactored termux_extract_dep_info to avoid re-sourcing build.sh.
Modified package extraction sequence to avoid storage hit and avoid spawning two more commands for checking if in-tar files are prefixed with `./` or not.

With these changes `termux_step_dependencies` execution time drops from `12s` to `7s` during `kf6-breeze-icons` rebuilding.

Also it reduces `~0.1s`-`~2.0s` per package during archive extracting step (depends on package size), which is good for packages with lots of dependencies (like `kf6-breeze-icons`).