@twaik 2025-05-01
This change makes it faster mostly for package rebuilding since I modified only this codepath.
But probably we can make regular building with dependency downloading much faster by parallelizing package downloads. I can take a look at this later.