@twaik 2025-05-01
Well, seems like `[[` is faster than `[`. Pure miracle.
... But not much.
```bash
$ time for ((i=0; i<1000000; i++)); do :; done
real    0m1.017s
user    0m1.017s
sys     0m0.000s

$ time for ((i=0; i<1000000; i++)); do [ -f /dev/null ]; done
real    0m1.949s
user    0m1.629s
sys     0m0.319s

$ time for ((i=0; i<1000000; i++)); do [[ -f /dev/null ]]; done
real    0m1.389s
user    0m1.075s
sys     0m0.314s
```