@Maxython 2025-05-07
> Ok, I'll merge it in 20 hours if nobody minds or after it gets 2+ more approves.

Wait, let me test how this all works with termux-pacman repos.